Bump milliHQ/download/npm from 1.1.0 to 2.0.0
When upgrading from a previous version of the module, Terraform may
fail with an error that the final plan has changed during apply.
That was caused by the milliHQ/download/npm module that downloaded the
source of the Lambda in parralel to running the apply.
In 2.0.0 this behavior was fixed, so that Terraform waits until the
download is finished before progreeding.